+++ modules/hostap/src/supp_api.h
@@ -8,6 +8,7 @@
 #ifndef ZEPHYR_SUPP_MGMT_H
 #define ZEPHYR_SUPP_MGMT_H
 
+#include <defs.h>
 #include <zephyr/net/wifi_mgmt.h>
 
 #ifndef MAX_SSID_LEN
@@ -19,6 +20,17 @@
 
 #define MAC_STR_LEN 18 /* for ':' or '-' separated MAC address string */
 #define CHAN_NUM_LEN 6 /* for space-separated channel numbers string */
+
+enum wifi_frequency_bands wpas_band_to_zephyr(enum wpa_radio_work_band band);
+
+enum wifi_wpa3_enterprise_type wpas_key_mgmt_to_zephyr_wpa3_ent(int key_mgmt);
+
+enum wifi_security_type wpas_key_mgmt_to_zephyr(bool is_hapd, void *config,
+						int key_mgmt, int proto, int pwe);
+
+const struct wifi_mgmt_ops *const get_wifi_mgmt_api(const struct device *dev);
+
+enum wifi_mfp_options get_mfp(enum mfp_options supp_mfp_option);
 
 /**
  * @brief Get version
@@ -194,6 +206,11 @@ int supplicant_mode(const struct device *dev, struct wifi_mode_info *mode);
 
 #if defined CONFIG_WIFI_NM_WPA_SUPPLICANT_CRYPTO_ENTERPRISE || \
 	defined CONFIG_WIFI_NM_HOSTAPD_CRYPTO_ENTERPRISE
+int is_eap_valid_security(int security);
+
+int process_cipher_config(struct wifi_connect_req_params *params,
+			  struct wifi_eap_cipher_config *cipher_config);
+
 /** Set Wi-Fi enterprise mode CA/client Cert and key
  *
  * @param dev Pointer to the device structure for the driver instance
@@ -289,50 +306,7 @@ int supplicant_get_wifi_conn_params(const struct device *dev,
 int supplicant_wps_config(const struct device *dev, struct wifi_wps_config_params *params);
 
 #ifdef CONFIG_AP
-#ifdef CONFIG_WIFI_NM_HOSTAPD_AP
-/**
- * @brief Get Wi-Fi AP Status
- *
- * @param dev Wi-Fi device
- * @param params AP status
- * @return 0 for OK; -1 for ERROR
- */
-int hapd_state(const struct device *dev, int *state);
-
-/**
- * @brief Wi-Fi AP configuration parameter.
- *
- * @param dev Wi-Fi device
- * @param params AP parameters
- * @return 0 for OK; -1 for ERROR
- */
-int supplicant_ap_config_params(const struct device *dev, struct wifi_ap_config_params *params);
-#else
-static inline int hapd_state(const struct device *dev, int *state)
-{
-	return -EINVAL;
-}
-#endif
-
-#ifdef CONFIG_WIFI_NM_HOSTAPD_WPS
-/** Start AP WPS PBC/PIN
- *
- * @param dev Pointer to the device structure for the driver instance
- * @param params wps operarion parameters
- *
- * @return 0 if ok, < 0 if error
- */
-int supplicant_ap_wps_config(const struct device *dev, struct wifi_wps_config_params *params);
-#endif
-
-/**
- * @brief Get Wi-Fi SAP status
- *
- * @param dev Wi-Fi device
- * @param status SAP status
- * @return 0 for OK; -1 for ERROR
- */
-int supplicant_ap_status(const struct device *dev, struct wifi_iface_status *status);
+int set_ap_bandwidth(const struct device *dev, enum wifi_frequency_bandwidths bandwidth);
 
 /**
  * @brief Set Wi-Fi AP configuration
@@ -363,6 +337,10 @@ int supplicant_ap_sta_disconnect(const struct device *dev,
 #endif /* CONFIG_AP */
 
 #ifdef CONFIG_WIFI_NM_WPA_SUPPLICANT_DPP
+#define SUPPLICANT_DPP_CMD_BUF_SIZE 384
+
+int dpp_params_to_cmd(struct wifi_dpp_params *params, char *cmd, size_t max_len);
+
 /**
  * @brief Dispatch DPP operations for STA
  *
@@ -371,16 +349,5 @@ int supplicant_ap_sta_disconnect(const struct device *dev,
  * @return 0 for OK; -1 for ERROR
  */
 int supplicant_dpp_dispatch(const struct device *dev, struct wifi_dpp_params *params);
-
-#ifdef CONFIG_WIFI_NM_HOSTAPD_AP
-/**
- * @brief Dispatch DPP operations for AP
- *
- * @param dev Wi-Fi interface name to use
- * @param dpp_params DPP action enum and params in string
- * @return 0 for OK; -1 for ERROR
- */
-int hapd_dpp_dispatch(const struct device *dev, struct wifi_dpp_params *params);
-#endif /* CONFIG_WIFI_NM_HOSTAPD_AP */
 #endif /* CONFIG_WIFI_NM_WPA_SUPPLICANT_DPP */
 #endif /* ZEPHYR_SUPP_MGMT_H */
