+++ modules/hostap/src/supp_main.h
@@ -55,14 +55,10 @@ struct wpa_global *zephyr_get_default_supplicant_context(void);
 struct wpa_supplicant *zephyr_get_handle_by_ifname(const char *ifname);
 
 #ifdef CONFIG_WIFI_NM_HOSTAPD_AP
-#include "common.h"
-#include "wpa_debug_zephyr.h"
-
-struct hostapd_iface *zephyr_get_hapd_handle_by_ifname(const char *ifname);
-void wpa_supplicant_msg_send(void *ctx, int level, enum wpa_msg_type type, const char *txt,
-			     size_t len);
-void hostapd_msg_send(void *ctx, int level, enum wpa_msg_type type, const char *buf, size_t len);
+struct hapd_interfaces *zephyr_get_default_hapd_context(void);
 #endif
+
+struct wpa_supplicant *zephyr_get_handle_by_ifname(const char *ifname);
 
 struct wpa_supplicant_event_msg {
 #ifdef CONFIG_WIFI_NM_HOSTAPD_AP
